Refactor SaleCoRequests components
