+++ src/components/hooks.js
@@ -1,0 +1,12 @@
+import { useState, useEffect } from 'react';
+
+export function useDebounce(value, delay) {
+  const [debounced, setDebounced] = useState(value);
+
+  useEffect(() => {
+    const handler = setTimeout(() => setDebounced(value), delay);
+    return () => clearTimeout(handler);
+  }, [value, delay]);
+
+  return debounced;
+}
